refactor: make use of tracker
